requirements.txt
virtualenv
flask
flask-cors
requests
sseclient
langchain
langchain_community
pinecone-client
openai
langchain_openai
pypdf